int id

+++ src/main/java/com/controlaltinsert/parkshark/division/domain/Division.java
@@ -29,11 +29,11 @@ public class Division {
         this.director = director;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
